requirements.txt
emoji==1.7.0
numpy==1.26.4
knockknock==0.1.7
transformers==4.36.2
Keras==2.15.0
tensorflow==2.15.0
neptune-client==0.9.19
torch==2.1.0
tqdm==4.66.2
pandas==2.1.4
GPUtil==1.4.0
gtrans==0.2
scikit-learn==1.8.0